CODERUB: Upgrade to .net 9

+++ Providus.XpressWallet.Core/Models/Services/Foundations/ExternalXpressWallet/ExternalWallet/ExternalAllWalletsResponse.cs
@@ -53,10 +53,10 @@ namespace Providus.XpressWallet.Core.Models.Services.Foundations.ExternalXpressW
             public string AccountNumber { get; set; }
 
             [JsonProperty("bookedBalance")]
-            public double BookedBalance { get; set; }
+            public decimal BookedBalance { get; set; }
 
             [JsonProperty("availableBalance")]
-            public double AvailableBalance { get; set; }
+            public decimal AvailableBalance { get; set; }
 
             [JsonProperty("accountReference")]
             public string AccountReference { get; set; }
